extracted teleoperation window to separate module
